refactor: 기존의 코드 infra 모듈에서 websocket 모듈로 이동
# Conflicts:
#	websocket/src/main/java/com/fintory/websocket/monitoring/config/StompStatsMetricsConfiguration.java
#	websocket/src/main/java/com/fintory/websocket/monitoring/config/WebSocketStatsConfig.java
