+++ websocket/src/main/java/com/fintory/websocket/config/WebSocketBrokerConfig.java
@@ -1,6 +1,5 @@
-package com.fintory.infra.domain.stock.config;
+package com.fintory.websocket.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
