+++ websocket/src/main/java/com/fintory/websocket/config/WebSocketInterceptor.java
@@ -1,4 +1,4 @@
-package com.fintory.infra.domain.stock.config;
+package com.fintory.websocket.config;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.server.ServerHttpRequest;
